Wait until the pbench agent daemonsets are ready
Although the wait time is usually negligible there is the potential to start a
test before the pbench agent daemonsets are up and running.  This p/r adds a
task to wait until the daemonsets are ready.  Additionally, there is a task
added to wait for scale-ci-tooling task deletion to complete.